requirements.txt
streamlit
google-genai
sqlite-utils
